WIO.End tests
