+++ workflows4s-core/src/main/scala/workflows4s/wio/builders/WIOBuilderMethods.scala
@@ -7,6 +7,8 @@ import workflows4s.wio.internal.WorkflowEmbedding
 
 trait WIOBuilderMethods[Ctx <: WorkflowContext] {
 
+  def end: WIO[Any, Nothing, Nothing, Ctx] = WIO.End[Ctx]()
+
   def getState[St <: WCState[Ctx]]: WIO[St, Nothing, St, Ctx] = WIO.Pure(s => s.asRight, WIO.Pure.Meta(NoError(), None))
 
   def embed[In, Err, Out <: WCState[InnerCtx], InnerCtx <: WorkflowContext, OS[_ <: WCState[InnerCtx]] <: WCState[Ctx]](
